# Backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_odatlar-bot_backend_internal_dto.BotUserResponse:
    properties:
      avatar_url:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_bot:
        type: boolean
      is_premium:
        type: boolean
      language_code:
        type: string
      last_name:
        type: string
      last_start_at:
        type: string
      level:
        type: integer
      level_up_xp:
        type: integer
      phone:
        type: string
      started_at:
        type: string
      telegram_id:
        type: integer
      username:
        type: string
      xp:
        type: integer
    type: object
...
  github_com_odatlar-bot_backend_internal_dto.CreateIndicatorRequest:
    properties:
      benefits:
        items:
          type: string
        maxItems: 20
        type: array
      name:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  github_com_odatlar-bot_backend_internal_dto.IndicatorHistoryEntryResponse:
    properties:
      completedAt:
        type: string
      date:
        type: string
      habitId:
        type: string
      habitName:
        type: string
      id:
        type: string
      isEmpty:
        type: boolean
      kind:
        type: string
      numericValue:
        type: number
      valueId:
        type: string
      valueLabel:
        type: string
    type: object
  github_com_odatlar-bot_backend_internal_dto.IndicatorHistoryResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.IndicatorHistoryEntryResponse'
        type: array
    type: object
  github_com_odatlar-bot_backend_internal_dto.IndicatorListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.IndicatorResponse'
        type: array
    type: object
  github_com_odatlar-bot_backend_internal_dto.IndicatorResponse:
    properties:
      benefits:
        items:
          type: string
        type: array
      completedToday:
        type: boolean
      createdAt:
        type: string
      id:
        type: string
      kind:
        type: string
      level:
        type: integer
      name:
        type: string
      streak:
        type: integer
      todayIndicatorValue:
        type: string
      xp:
        type: integer
      xp_reward:
        type: integer
    type: object
  github_com_odatlar-bot_backend_internal_dto.LeaderboardEntryResponse:
    properties:
      is_me:
        type: boolean
      level:
        type: integer
      name:
        type: string
      rank:
        type: integer
      xp:
        type: integer
    type: object
  github_com_odatlar-bot_backend_internal_dto.LeaderboardResponse:
    properties:
      current_user_rank:
        type: integer
      data:
        items:
          $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.LeaderboardEntryResponse'
        type: array
    type: object
  github_com_odatlar-bot_backend_internal_dto.LogIndicatorRequest:
    properties:
      is_empty:
        type: boolean
      value_id:
        maxLength: 100
        type: string
      value_label:
        maxLength: 255
        type: string
    required:
    - value_id
    - value_label
    type: object
...
  github_com_odatlar-bot_backend_internal_dto.MonitoredDominant:
    properties:
      created_at:
        type: string
      id:
        type: string
      last_session_at:
        type: string
      sessions_completed:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
  github_com_odatlar-bot_backend_internal_dto.MonitoredIndicator:
    properties:
      created_at:
        type: string
      id:
        type: string
      last_logged_at:
        type: string
      last_value_label:
        type: string
      logs_total:
        type: integer
      name:
        type: string
    type: object
  github_com_odatlar-bot_backend_internal_dto.MonitoredPractice:
    properties:
      completions_total:
        type: integer
      created_at:
        type: string
      id:
        type: string
      last_completed_at:
        type: string
      name:
        type: string
      streak:
        type: integer
    type: object
...
  github_com_odatlar-bot_backend_internal_dto.UpdateIndicatorRequest:
    properties:
      benefits:
        items:
          type: string
        maxItems: 20
        type: array
      name:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  github_com_odatlar-bot_backend_internal_dto.UserActivityEntry:
    properties:
      at:
        type: string
      detail:
        type: string
      kind:
        description: practice_complete | indicator_log | dominant_session
        type: string
      title:
        type: string
    type: object
  github_com_odatlar-bot_backend_internal_dto.UserActivityResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.UserActivityEntry'
        type: array
      from:
        type: string
      limit:
        type: integer
      to:
        type: string
    type: object
  github_com_odatlar-bot_backend_internal_dto.UserMonitoringResponse:
    properties:
      dominants:
        items:
          $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.MonitoredDominant'
        type: array
      generated_at:
        type: string
      indicators:
        items:
          $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.MonitoredIndicator'
        type: array
      practices:
        items:
          $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.MonitoredPractice'
        type: array
      recent_activity:
        items:
          $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.UserActivityEntry'
        type: array
      summary:
        $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.UserMonitoringSummary'
      user:
        $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.BotUserResponse'
    type: object
  github_com_odatlar-bot_backend_internal_dto.UserMonitoringSummary:
    properties:
      best_streak:
        type: integer
      days_since_start:
        type: integer
      dominant_count:
        type: integer
      dominant_sessions_today:
        type: integer
      dominant_sessions_total:
        type: integer
      dominant_sessions_week:
        type: integer
      indicator_count:
        type: integer
      indicator_logs_today:
        type: integer
      indicator_logs_total:
        type: integer
      indicator_logs_week:
        type: integer
      level:
        type: integer
      level_up_xp:
        type: integer
      practice_completions_today:
        type: integer
      practice_completions_total:
        type: integer
      practice_completions_week:
        type: integer
      practice_count:
        type: integer
      rank:
        type: integer
      xp:
        type: integer
    type: object
...
  github_com_odatlar-bot_backend_pkg_response.ErrorBody:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@odatlar-bot.uz
    name: API Support
  description: Admin panel, Bot Runtime (Telegram WebApp), Qo'llanma (videolar, kurslar,
    fayllar) va Bildirishnomalar APIlari.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Odatlar Bot API
  version: "1.0"
paths:
...
  /bot-runtime/indicators:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.IndicatorListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
      security:
      - TelegramInitData: []
      summary: Ko'rsatkichlar ro'yxati
      tags:
      - Indicators
    post:
      consumes:
      - application/json
      parameters:
      - description: Ko'rsatkich ma'lumotlari
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.CreateIndicatorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.IndicatorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
      security:
      - TelegramInitData: []
      summary: Yangi ko'rsatkich yaratish
      tags:
      - Indicators
  /bot-runtime/indicators/{id}:
    delete:
      parameters:
      - description: Ko'rsatkich ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
      security:
      - TelegramInitData: []
      summary: Ko'rsatkichni o'chirish
      tags:
      - Indicators
    put:
      consumes:
      - application/json
      parameters:
      - description: Ko'rsatkich ID
        in: path
        name: id
        required: true
        type: integer
      - description: Yangilangan ma'lumotlar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.UpdateIndicatorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.IndicatorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
      security:
      - TelegramInitData: []
      summary: Ko'rsatkichni yangilash
      tags:
      - Indicators
  /bot-runtime/indicators/{id}/log:
    post:
      consumes:
      - application/json
      parameters:
      - description: Ko'rsatkich ID
        in: path
        name: id
        required: true
        type: integer
      - description: Qiymat ma'lumotlari
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.LogIndicatorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.IndicatorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
      security:
      - TelegramInitData: []
      summary: Ko'rsatkich qiymatini yozish
      tags:
      - Indicators
  /bot-runtime/indicators/history:
    get:
      parameters:
      - description: Boshlanish sanasi (YYYY-MM-DD)
        in: query
        name: from
        required: true
        type: string
      - description: Tugash sanasi (YYYY-MM-DD)
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.IndicatorHistoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
      security:
      - TelegramInitData: []
      summary: Ko'rsatkichlar tarixi
      tags:
      - Indicators
  /bot-runtime/leaderboard:
    get:
      parameters:
      - description: Natijalar soni (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.LeaderboardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
      security:
      - TelegramInitData: []
      summary: Reyting jadvali (WebApp)
      tags:
      - Leaderboard
...
  /bot/users/{id}/activity:
    get:
      description: Amaliyot bajarilishlari, indikator kiritishlari va dominanta sessiyalarini
        xronologik tartibda qaytaradi
      parameters:
      - description: Bot user ID
        in: path
        name: id
        required: true
        type: integer
      - description: Boshlanish sanasi (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: Tugash sanasi (YYYY-MM-DD)
        in: query
        name: to
        type: string
      - default: 50
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.UserActivityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Foydalanuvchi faoliyat lentasi
      tags:
      - Bot Monitoring
  /bot/users/{id}/monitoring:
    get:
      description: Bitta foydalanuvchi bo'yicha profil, umumiy ko'rsatkichlar, amaliyot/indikator/dominanta
        tafsilotlari va so'nggi faoliyat lentasi
      parameters:
      - description: Bot user ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.UserMonitoringResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Foydalanuvchini to'liq monitoring
      tags:
      - Bot Monitoring
...
securityDefinitions:
  BearerAuth:
    description: 'JWT token. Format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
  BotToken:
    description: 'Telegram bot token. Format: Bearer {bot_token}'
    in: header
    name: Authorization
    type: apiKey
  TelegramInitData:
    description: 'Telegram WebApp initData. Alternativa: Authorization: tma {initData}'
    in: header
    name: X-Telegram-Init-Data
    type: apiKey
swagger: "2.0"