# Backend/internal/handler/leaderboard_handler.go
package handler

import (
	"errors"
	"strconv"

	"github.com/gin-gonic/gin"
	"github.com/odatlar-bot/backend/internal/dto"
	"github.com/odatlar-bot/backend/internal/service"
	"github.com/odatlar-bot/backend/pkg/response"
)

type LeaderboardHandler struct {
	service *service.LeaderboardService
}

func NewLeaderboardHandler(svc *service.LeaderboardService) *LeaderboardHandler {
	return &LeaderboardHandler{service: svc}
}

var _ dto.LeaderboardResponse

// Get godoc
// @Summary      Reyting jadvali (WebApp)
// @Tags         Leaderboard
// @Produce      json
// @Security     TelegramInitData
// @Param        limit  query  int  false  "Natijalar soni (default 20)"
// @Success      200    {object}  dto.LeaderboardResponse
// @Failure      401    {object}  response.ErrorBody
// @Failure      404    {object}  response.ErrorBody
// @Failure      500    {object}  response.ErrorBody
// @Router       /bot-runtime/leaderboard [get]
func (h *LeaderboardHandler) Get(c *gin.Context) {
	tgUser := getTelegramUser(c)
	if tgUser == nil {
		return
	}

	limit := 20
	if raw := c.Query("limit"); raw != "" {
		if parsed, err := strconv.Atoi(raw); err == nil {
			limit = parsed
		}
	}

	result, err := h.service.Get(c.Request.Context(), tgUser, limit)
	if err != nil {
		handleLeaderboardError(c, err)
		return
	}
	response.OK(c, result)
}

func handleLeaderboardError(c *gin.Context, err error) {
	switch {
	case errors.Is(err, service.ErrRuntimeBotUserNotFound):
		response.NotFound(c, "bot user not found")
	default:
		response.InternalError(c, "leaderboard request failed", err)
	}
}